+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (GH #502) Items being removed from IndexedDB on network errors.
 - (GH #514) Fixed container and metadata updates
+- (GH #547) Fixed upload button being enabled when there are no files to upload.
 
 ## [v2.0.1]